// src/app/types.d.ts
export interface ImageInfo {
  id?: string;
  url?: string;
  type?: string;
  createdAt?: string;
  compressedImg?: string;
  fileName?: string;
  originalSize?: number;
}

export interface Price {
  value: number;
  currency: string;
}

export interface Coordinates {
  longitude: number;
  latitude: number;
}

export interface Photo {
  id: string;
  orderNumber: number;
  image: ImageInfo;
  type: string;
  isCover: boolean;
}